Remove stdlib and runtime dependencies on Foundation and CF
