+++ include/swift/Runtime/ObjCBridge.h
@@ -79,6 +79,11 @@ namespace swift {
 SWIFT_RUNTIME_EXPORT
 void swift_rootObjCDealloc(HeapObject *self);
 
+// Uses Swift bridging to box a C string into an NSString without introducing
+// a link-time dependency on NSString.
+SWIFT_CC(swift) SWIFT_RUNTIME_STDLIB_API
+id swift_stdlib_NSStringFromUTF8(const char *cstr, int len);
+
 }
 
 #endif // SWIFT_OBJC_INTEROP
