+++ stdlib/public/runtime/SwiftObject.mm
@@ -173,25 +173,26 @@ Class _swift_classOfObjCHeapObject(OpaqueValue *value) {
 }
 
 SWIFT_CC(swift) SWIFT_RUNTIME_STDLIB_API
-NSString *swift_stdlib_getDescription(OpaqueValue *value,
+id swift_stdlib_getDescription(OpaqueValue *value,
                                       const Metadata *type);
 
-NSString *swift::getDescription(OpaqueValue *value, const Metadata *type) {
-  auto result = swift_stdlib_getDescription(value, type);
+id swift::getDescription(OpaqueValue *value, const Metadata *type) {
+  id result = swift_stdlib_getDescription(value, type);
   type->vw_destroy(value);
   return [result autorelease];
 }
 
-static NSString *_getObjectDescription(SwiftObject *obj) {
+static id _getObjectDescription(SwiftObject *obj) {
   swift_retain((HeapObject*)obj);
   return getDescription((OpaqueValue*)&obj,
                         _swift_getClassOfAllocated(obj));
 }
 
-static NSString *_getClassDescription(Class cls) {
-  return NSStringFromClass(cls);
+static id _getClassDescription(Class cls) {
+  const char *name = class_getName(cls);
+  int len = strlen(name);
+  return [swift_stdlib_NSStringFromUTF8(name, len) autorelease];
 }
-
 
 @implementation SwiftObject
 + (void)initialize {
@@ -414,21 +415,21 @@ static NSString *_getClassDescription(Class cls) {
                                                                  object2);
 }
 
-- (NSString *)description {
+- (id /* NSString */)description {
   return _getObjectDescription(self);
 }
-- (NSString *)debugDescription {
+- (id /* NSString */)debugDescription {
   return _getObjectDescription(self);
 }
 
-+ (NSString *)description {
++ (id /* NSString */)description {
   return _getClassDescription(self);
 }
-+ (NSString *)debugDescription {
++ (id /* NSString */)debugDescription {
   return _getClassDescription(self);
 }
 
-- (NSString *)_copyDescription {
+- (id /* NSString */)_copyDescription {
   // The NSObject version of this pushes an autoreleasepool in case -description
   // autoreleases, but we're OK with leaking things if we're at the top level
   // of the main thread with no autorelease pool.
@@ -436,15 +437,7 @@ static NSString *_getClassDescription(Class cls) {
 }
 
 - (CFTypeID)_cfTypeID {
-  // Adopt the same CFTypeID as NSObject.
-  static CFTypeID result;
-  static dispatch_once_t predicate;
-  dispatch_once_f(&predicate, &result, [](void *resultAddr) {
-    id obj = [[NSObject alloc] init];
-    *(CFTypeID*)resultAddr = [obj _cfTypeID];
-    [obj release];
-  });
-  return result;
+  return (CFTypeID)1; //NSObject's CFTypeID is constant
 }
 
 // Foundation collections expect these to be implemented.
