+++ stdlib/public/runtime/SwiftValue.mm
@@ -336,7 +336,7 @@ swift::findSwiftValueConformances(const ExistentialTypeMetadata *existentialType
       selfHeader->type, hashableConformance);
 }
 
-static NSString *getValueDescription(__SwiftValue *self) {
+static id getValueDescription(__SwiftValue *self) {
   const Metadata *type;
   const OpaqueValue *value;
   std::tie(type, value) = getValueFromSwiftValue(self);
@@ -346,15 +346,15 @@ static NSString *getValueDescription(__SwiftValue *self) {
   auto copy = type->allocateBufferIn(&copyBuf);
   type->vw_initializeWithCopy(copy, const_cast<OpaqueValue *>(value));
 
-  NSString *string = getDescription(copy, type);
+  id string = getDescription(copy, type);
   type->deallocateBufferIn(&copyBuf);
   return string;
 }
 
-- (NSString *)description {
+- (id /* NSString */)description {
   return getValueDescription(self);
 }
-- (NSString *)debugDescription {
+- (id /* NSString */)debugDescription {
   return getValueDescription(self);
 }
 
@@ -363,11 +363,11 @@ static NSString *getValueDescription(__SwiftValue *self) {
 - (const Metadata *)_swiftTypeMetadata {
   return getSwiftValueTypeMetadata(self);
 }
-- (NSString *)_swiftTypeName {
+- (id /* NSString */)_swiftTypeName {
   TypeNamePair typeName
     = swift_getTypeName(getSwiftValueTypeMetadata(self), true);
-
-  return [NSString stringWithUTF8String: typeName.data];
+  id str = swift_stdlib_NSStringFromUTF8(typeName.data, typeName.length);
+  return [str autorelease];
 }
 - (const OpaqueValue *)_swiftValue {
   return getValueFromSwiftValue(self).second;
